requirements.txt
requests
pyttsx
SpeechRecognition